--- ui-sdk-components-sample/app/src/main/java/com/karhoo/samples/networksdk/base/state/NoObserverAttachedException.kt
@@ -1,3 +1,0 @@
-package com.karhoo.samples.networksdk.base.state
-
-class NoObserverAttachedException(message: String) : Exception(message)
