--- ui-sdk-components-sample/app/src/main/java/com/karhoo/samples/networksdk/base/state/ViewModelContract.kt
@@ -1,6 +1,0 @@
-package com.karhoo.samples.networksdk.base.state
-
-// A contract to have predefined events to update the state of the widget.
-internal interface ViewModelContract<EVENT> {
-    fun process(viewEvent: EVENT)
-}
