--- ui-sdk-components-sample/app/src/main/java/com/karhoo/samples/networksdk/config/ConfigContract.kt
@@ -1,10 +1,0 @@
-package com.karhoo.samples.networksdk.config
-
-import com.karhoo.uisdk.KarhooUISDKConfiguration
-
-class ConfigContract {
-
-    interface Module {
-        fun karhooUserConfiguration(): KarhooUISDKConfiguration
-    }
-}
