--- ui-sdk-components-sample/app/src/main/java/com/karhoo/samples/networksdk/configuration/ConfigurationStatus.kt
@@ -1,3 +1,0 @@
-package com.karhoo.samples.networksdk.configuration
-
-data class ConfigurationStatus(var signedIn: Boolean)
