+++ ui-sdk-components-sample/app/src/main/java/com/karhoo/samples/uisdk/components/config/ConfigContract.kt
@@ -1,0 +1,10 @@
+package com.karhoo.samples.uisdk.components.config
+
+import com.karhoo.uisdk.KarhooUISDKConfiguration
+
+class ConfigContract {
+
+    interface Module {
+        fun karhooUserConfiguration(): KarhooUISDKConfiguration
+    }
+}
