+++ ui-sdk-components-sample/app/src/main/java/com/karhoo/samples/uisdk/components/configuration/ConfigurationStatus.kt
@@ -1,0 +1,3 @@
+package com.karhoo.samples.uisdk.components.configuration
+
+data class ConfigurationStatus(var signedIn: Boolean)
