+++ ui-sdk-components-sample/app/src/main/res/layout/fragment_trip_booking.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/blue_green"
     android:padding="@dimen/padding_default"
-    tools:context=".booking.TripBookingFragment">
+    tools:context="com.karhoo.samples.uisdk.components.booking.TripBookingFragment">
 
     <com.karhoo.uisdk.screen.booking.booking.bookingrequest.BookingRequestView
         android:id="@+id/booking_request_widget"
