+++ ui-sdk-components-sample/app/src/main/res/layout/fragment_trip_quotes.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/sky_blue"
     android:padding="@dimen/padding_default"
-    tools:context=".quotes.TripQuotesFragment">
+    tools:context="com.karhoo.samples.uisdk.components.quotes.TripQuotesFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/quotes_layout"
