--- ui-sdk-components-sample/app/src/test/java/com/karhoo/samples/networksdk/ExampleUnitTest.kt
@@ -1,16 +1,0 @@
-package com.karhoo.samples.networksdk
-
-import org.junit.Assert.assertEquals
-import org.junit.Test
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-class ExampleUnitTest {
-    @Test
-    fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
-    }
-}
